refactor(account): move TokenList to separate file
